Making a better client
Now it's time to make a client that will request those offers for the
web thing.

It will receive all the necessary parameters to make a successful
request, add that special hashkey and done.

The client class responsibility will be to deal with the params and
delegate the rest, mainly coordinating stuff. The others are
self-explanatory, I think...