HTTP with Curb
Changed the default client to Curb and abstracted its interface with an adapter class.

This class is a simple object that will be instantiated and then called on get with the url to make the request.

With Curb there's the benefit of keep-alive connections without being constrained by a block.